requirements.txt
PIL
python-telegram-bot
facebook-sdk
google-genai
moondream
psycopg2-binary
python-dotenv
